Minor project-wide refactoring.
Code Style:

- Add missing final's.
- Move private field to top of class.
- Remove unnecessary static class qualifiers.
- Use == for comparing Enum's, instead of .equals
- Replace i++ (post-increment) with ++i (pre-increment).
- Make non-static final field camelcase instead of uppercase with underscores.
- Explicitly initialize fields with default values where they were used without explicit initialization.

Accuracy & Readability:

- Fix spelling errors.
- Remove unnecessary casts.
- Simplify if statements where possible.
- Simplify equals & hashCode methods where possible.
- Testing: Use assertEquals instead of assertTrue where possible.
- Make overridden method parameter names match the original parameter names.
- Testing: Make expected value 1st parameter, actual 2nd for junit assert* methods.
- Replace float/double to int casting with Math.round, which avoids off-by-one errors.

Performance:

- Make fields static where possible.
- Make methods static where possible.
- Use EnumSet & EnumMap where possible.
- Remove unnecessary null checks when using instanceof.
- Iterate over Map's EntrySet in for loops where possible.
- Logging: Use slf4j {} construct (fast) instead of string concatenation (slow).

+++ client/src/main/java/com/forerunnergames/peril/client/ui/screens/LibGdxScreenSize.java
@@ -133,15 +133,15 @@ public final class LibGdxScreenSize implements ScreenSize
 
     if (currentWidth != previousWidth)
     {
-      actualToReferenceScaling.x = (float) GraphicsSettings.REFERENCE_SCREEN_WIDTH / (float) currentWidth;
-      referenceToActualScaling.x = (float) currentWidth / (float) GraphicsSettings.REFERENCE_SCREEN_WIDTH;
+      actualToReferenceScaling.x = GraphicsSettings.REFERENCE_SCREEN_WIDTH / (float) currentWidth;
+      referenceToActualScaling.x = currentWidth / (float) GraphicsSettings.REFERENCE_SCREEN_WIDTH;
       previousWidth = currentWidth;
     }
 
     if (currentHeight != previousHeight)
     {
-      actualToReferenceScaling.y = (float) GraphicsSettings.REFERENCE_SCREEN_HEIGHT / (float) currentHeight;
-      referenceToActualScaling.y = (float) currentHeight / (float) GraphicsSettings.REFERENCE_SCREEN_HEIGHT;
+      actualToReferenceScaling.y = GraphicsSettings.REFERENCE_SCREEN_HEIGHT / (float) currentHeight;
+      referenceToActualScaling.y = currentHeight / (float) GraphicsSettings.REFERENCE_SCREEN_HEIGHT;
       previousHeight = currentHeight;
     }
   }
